+++ CHANGELOG.md
@@ -10,6 +10,8 @@ For examples and guidelines see [https://keepachangelog.com/](https://keepachang
 
 - Fix ordering of syslog message to comply with RFC5424
 
+
+
 ## [Released]